refactor: API 호출 경로 변경
