+++ src/pages/explore/index.tsx
@@ -8,7 +8,7 @@ import { ListView } from "./_components/layout/ListView";
 import { GridView } from "./_components/layout/GridView";
 import { AppTabs } from "../../components/common/AppTabs";
 import { cn } from "../../lib/utils";
-import { fetchDevelopers } from "../../lib/api";
+import { fetchDevelopers } from "../../lib/api/developers";
 
 export const ExplorePage = () => {
   const [activeTab, setActiveTab] = useState<ExploreTabValue>("nearby");
